Fix 404 error when not foun

+++ src/controllers/unit.controller.ts
@@ -44,13 +44,18 @@ export class UnitController {
       res.status(200).json( await this.unitService.findAll());
   }
   async delete(req: Request, res: Response) {
+    if ( await this.unitService.findById(req.params.id) == undefined ) {
+      res.status(HttpStatusCode.NOT_FOUND).json({});
+    }
+    else {
       const unitDeleteStatus: Boolean = await this.unitService.delete(req.params.id);
       const relationsDeleteStatus: Boolean = await this.relationController.deleteByConexion(req.params.id);
       if ( unitDeleteStatus === true && relationsDeleteStatus === true) {
-        res.status(HttpStatusCode.NO_CONTENT);
+        res.status(HttpStatusCode.NO_CONTENT).json({});
       }
       else {
-        res.status(HttpStatusCode.INTERNAL_SERVER_ERROR);
+        res.status(HttpStatusCode.INTERNAL_SERVER_ERROR).json({});
       }
+    }
   }
 }
